Update auth.py
Firstly, a function should only be responsable for a single task, so I took out the account information and made that into its own function. I then simplified the syntax in this new function so that it is much easier to read.

Also I combinded the Username and Password look up into only one if statement to reduce nesting.

I would call this 2 fixes.